Back to boost/unordered_map -- 2s
The bottleneck in this code is the hash function. Commenting out the map
step reduces the runtime 20 fold, i.e. 95% of the time is spent in the
hash function.

For some reason, boost::unordered_map is twice as fast as the
std::unordered_map. I don't like the extra dependency.

It might be worthwhile to implement my own hash function specialized for
biological sequence.