Add C++11 flag
I tried several methods.

First, this in the cpp headers:
```
// [[Rcpp::plugins(cpp11)]]
```

But no `-std=c++11` flag appeared.

Next I tried setting the flag in the R shell

```
> Sys.setenv("PKG_CXXFLAGS"="-std=c++11")
> devtools::clean_dll()
> devtools::install()
```

This worked, but of course is not an acceptable solution, since I can't
enter this into the shell by hand with every installation.

So I tried adding the `Sys.setenv` command in the R code itself, but
this did not result in a flag being added. Presumably because the
compilation actually occurs before this code is executed. It would
behoove me to study the exact order of events.

Finally I tried adding the flag to the `src/Makevars` file. This finally
worked.

```
PKG_CXXFLAGS=-std=c++11
```

Actually, using my compiler, g++ v6.2.1, c11 is default, so the flag isn't
necessary. But of course, Travis, as always, is a total pain. It uses an
ancient compiler, so will need some work.